improved user controllers and model

+++ src/controllers/users.controller.js
@@ -1,9 +1,12 @@
 const bcrypt = require('bcrypt');
 
 const { Op } = require('sequelize');
+
 const {
 	models: { User, Post, Favorites, Reactions },
 } = require('../sequelize');
+
+const { parseModel } = require('../helpers/sequelize');
 
 module.exports = {
 	async getUsers(req, res, next) {
@@ -12,12 +15,13 @@ module.exports = {
 		if (!email) email = '';
 		if (!username) username = '';
 		try {
-			const users = await User.findAll({
+			let users = await User.findAll({
 				where: {
 					name: { [Op.iLike]: `%${name}%` },
 					email: { [Op.iLike]: `%${email}%` },
 					username: { [Op.iLike]: `%${username}%` },
 				},
+				attributes: ['id', 'name', 'email', 'username'],
 				include: [
 					{
 						model: Post,
@@ -33,6 +37,20 @@ module.exports = {
 					},
 				],
 			});
+
+			if (!users || !users.length) {
+				return res
+					.status(404)
+					.json({ error: true, message: 'No users found.' });
+			}
+
+			users = users.map((user) => parseModel(user));
+			users = users.map((user) => ({
+				...user,
+				favorites: user.favorites
+					.map((fav) => parseModel(fav))
+					.map((fav) => fav.postId),
+			}));
 			res.status(200).json(users);
 		} catch (e) {
 			next(e);
@@ -41,7 +59,14 @@ module.exports = {
 	async getUser(req, res, next) {
 		const { id } = req.params;
 		try {
+			if (!id || id === '') {
+				return res
+					.status(400)
+					.json({ error: true, message: 'ID must be provided.' });
+			}
+
 			const user = await User.findByPk(id, {
+				attributes: ['id', 'name', 'email', 'username'],
 				include: [
 					{
 						model: Post,
@@ -57,6 +82,9 @@ module.exports = {
 					},
 				],
 			});
+			if (!user) {
+				return res.status(404).json({ error: true, message: 'No user found.' });
+			}
 			res.status(200).json(user);
 		} catch (e) {
 			next(e);
@@ -66,6 +94,12 @@ module.exports = {
 		const { username, email, name } = req.body;
 		let { password } = req.body;
 		try {
+			if (!username || !email || !name || !password) {
+				return res
+					.status(400)
+					.json({ error: true, message: 'Every field must be provided.' });
+			}
+
 			bcrypt.genSalt(10, async (err, salt) => {
 				if (err) return res.send(err);
 				bcrypt.hash(password, salt, async (err, hash) => {
@@ -87,23 +121,52 @@ module.exports = {
 	async updateUser(req, res, next) {
 		const { id } = req.params;
 		const { username, email, name } = req.body;
+		const userId = req.user.id;
 		try {
-			const userModified = await User.update(
+			if (userId !== id) {
+				return res
+					.status(401)
+					.json({ error: true, message: 'User cannot update other users.' });
+			}
+			const found = await User.findByPk(id);
+
+			if (!found) {
+				return res
+					.status(404)
+					.json({ error: true, message: 'User does not exist.' });
+			}
+			await User.update(
 				{ username, email, name },
 				{ where: { id: { [Op.eq]: id } } }
 			);
-			res.status(200).json(userModified);
+
+			res.status(200).json({ message: 'User updated successfully!' });
 		} catch (e) {
 			next(e);
 		}
 	},
 	async deleteUser(req, res, next) {
 		const { id } = req.params;
+		const userId = req.user.id;
 		try {
-			const userModified = await User.destroy({
+			if (userId !== id) {
+				return res
+					.status(401)
+					.json({ error: true, message: 'User cannot delete other users.' });
+			}
+
+			const found = await User.findByPk(id);
+			if (!found) {
+				return res
+					.status(404)
+					.json({ error: true, message: 'User does not exist.' });
+			}
+
+			await User.destroy({
 				where: { id: { [Op.eq]: id } },
 			});
-			res.status(204).json(userModified);
+
+			res.status(204).json({ message: 'User deleted successfully!' });
 		} catch (e) {
 			next(e);
 		}
